requirements.txt
speechrecognition
google-generativeai
pyttsx3
python-dotenv
streamlit
pathlib2
pyaudio